% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{info}
\alias{info}
\title{Carga de la informacion del mapa
El \code{read.csv} Permite cargar el archivo .csv  para cargar la informacion del mapa}
\format{An object of class \code{data.frame} with 338 rows and 17 columns.}
\usage{
info
}
\arguments{
\item{../Data/VEN_adm2.csv}{Fichero donde esta ubicado el archivo}

\item{header}{Es la cabecera del archivo, es decir, los nombres de las variables}

\item{sep}{Es el metodo para separar los datos en el archivo}
}
\description{
Carga de la informacion del mapa
El \code{read.csv} Permite cargar el archivo .csv  para cargar la informacion del mapa
}
\keyword{datasets}
