% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{map}
\alias{map}
\title{Carga del archivo .shp del mapa de venezuela
El \code{readOGR} Permite cargar el archivo .shp  para cargar el mapa}
\format{An object of class \code{SpatialPolygonsDataFrame} with 338 rows and 14 columns.}
\usage{
map
}
\arguments{
\item{VEN_adm2.shp}{Fichero donde esta ubicado el archivo}

\item{layer}{Es la capa del mapa}
}
\value{
Me devuelve el mapa
}
\description{
Carga del archivo .shp del mapa de venezuela
El \code{readOGR} Permite cargar el archivo .shp  para cargar el mapa
}
\keyword{datasets}
