% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Archivo Servidor}
\usage{
server(input, output)
}
\arguments{
\item{eventReactive}{es la funcion que establece que el evento del selectInput sea reactivo}

\item{input$entradas}{es la informacion del selectInput
 
Dato reactivo \code{datosVInput}
 
Permite crear un evento reactivo con la carga viral, celulas CD4 y CD8 que el usuario desee visualizar
Permite una visualizacion dinamica}

\item{eventReactive}{es la funcion que establece que el evento del selectInput sea reactivo}

\item{input$DatosV}{es la informacion del selectInput
 
Dato reactivo \code{selectajusteInput}
Permite crear un evento reactivo con la carga viral, celulas CD4 y CD8 que el usuario desee visualizar
Permite una visualizacion dinamica}

\item{eventReactive}{es la funcion que establece que el evento del selectInput sea reactivo}

\item{input$selectajuste}{es la informacion del selectInput

Dato reactivo \code{selectcargaInput}
Permite crear un evento reactivo con la carga viral, celulas CD4 y CD8 que el usuario desee visualizar
Permite una visualizacion dinamica}

\item{eventReactive}{es la funcion que establece que el evento del selectInput sea reactivo}

\item{input$selectcarga}{es la informacion del selectInput

Dato reactivo \code{dato}
Permite crear un evento reactivo con la informacion cargada del archivo
Permite una visualizacion dinamica}

\item{reactive}{es la funcion que establece que la informacion sea reactiva
El \code{file1} Carga el archivo
El \code{input$file} es la informacion del archivo que se cargo
\code{is.null(file1)} se realiza una condicion para chequear que realmente se carga el archivo
si no cargo nada solo devuelve un \code{return()}
en caso contrario, me guarda la informacion
en \code{read.table}}

\item{file1$datapath}{es la ubicacion del archivo}

\item{sep}{es el metodo de separacion de todo el archivo}

\item{header}{verifica si tiene cabecera el archivo}
}
\description{
Archivo donde se realiza el calculo, graficos y mapa de la aplicacion
}
\author{
Stephanie Correa

Dato reactivo \code{entradasInput} 
Permite crear un evento reactivo con el periodo que el usuario desee visualizar en la categoria por genero
Permite una visualizacion dinamica
}
