% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Dato reactivo \code{entradasInput}
Permite crear un evento reactivo con el periodo que el usuario desee visualizar en la categoria por genero
Dato reactivo \code{datosVInput}}
\usage{
server(input, output)
}
\description{
Crea un \code{renderTable} con la informacion del nombre, la extension y la ubicacion del archivo
Verifica si esta cargado el archivo, sino se regresa
si carga muestra la informacion
Muestra la informacion del archivo

Crea un \code{renderDataTable} con la informacion cargada del archivo
Primero verifica si se carga el archivo
Muestra el resumen de la informacion con las variables cargadas del sistema
Me carga el panel de la informacion del archivo

Cuando la informacion es cargada al sistema, aparece un pequeño panel con las opciones
de acerca del archivo, informacion y resumen de la informacion
Visualizacion del mapa
Crea el renderLeaflet para visualizar el mapa
Muestra la visualizacion grafica de la exploracion de datos
Muestra un grafico de tortas con el genero
Muestra un histograma de los edades de los datos cargados
Muestra un grafico de lineas con el comportamiento de la carga viral plasmatica, celulas CD4 y CD8
Muestra el resultado de la aplicacion del modelo lineal mixto
Muestra la salida del formula del modelo lineal mixto
Muestra los residuos encontrados en la formaula del modelo lineal mixto
Graficacion de los residuos del modelo lineal mixto
Genera el reporte con los datos
}
\details{
Permite crear un evento reactivo con la carga viral, celulas CD4 y CD8 que el usuario desee visualizar
Dato reactivo selectajusteInput
Permite crear un evento reactivo con la carga viral, celulas CD4 y CD8 que el usuario desee visualizar
Dato reactivo \code{selectcargaInput}
Permite crear un evento reactivo con la carga viral, celulas CD4 y CD8 que el usuario desee visualizar
Dato reactivo \code{dato}
Permite crear un evento reactivo con la informacion cargada del archivo
Dato Reactivo extension
Muestra la informacion referente al nombre del archivo y la extension
}
